+++ instrumentation/opentelemetry-instrumentation-requests/src/opentelemetry/instrumentation/requests/__init__.py
@@ -65,10 +65,7 @@ _SUPPRESS_HTTP_INSTRUMENTATION_KEY = context.create_key(
 # pylint: disable=unused-argument
 # pylint: disable=R0915
 def _instrument(
-    tracer, 
-    span_callback=None, 
-    name_callback=None, 
-    tracked_url_callback=None,
+    tracer, span_callback=None, name_callback=None, tracked_url_callback=None
 ):
     """Enables tracing of all requests calls that go through
     :code:`requests.session.Session.request` (this includes