+++ instrumentation/opentelemetry-instrumentation-requests/tests/test_requests_integration.py
@@ -130,11 +130,14 @@ class RequestsIntegrationTestBase(abc.ABC):
         httpretty.register_uri(
             httpretty.GET, url_404, status=404,
         )
+
         def tracked_url_callback(method, url):
             return "404" not in url
 
         RequestsInstrumentor().uninstrument()
-        RequestsInstrumentor().instrument(tracked_url_callback=tracked_url_callback)
+        RequestsInstrumentor().instrument(
+            tracked_url_callback=tracked_url_callback
+        )
         self.perform_request(self.URL)
         self.perform_request(url_404)
 
@@ -271,7 +274,7 @@ class RequestsIntegrationTestBase(abc.ABC):
             {
                 SpanAttributes.HTTP_METHOD: "GET",
                 SpanAttributes.HTTP_URL: self.URL,
-                SpanAttributes.HTTP_STATUS_CODE: 200, 
+                SpanAttributes.HTTP_STATUS_CODE: 200,
                 "http.response.body": "Hello!",
             },
         )